Out styling changes

+++ src/main/java/ClientTypes/TCPClient.java
@@ -91,7 +91,7 @@ public class TCPClient extends Client {
         }
 
         //calculate and print throughput
-        double timeInSecs = (System.nanoTime() - startTime)/1E9;
+        double timeInSecs = ((double) (System.nanoTime() - startTime))/1E9;
         double throughput = (numOfMsgs * msg.length * 8)/timeInSecs;
         System.out.println(numOfMsgs + "X" + msg.length + "B: " + throughput + " b/s");
